social_media_dashboard/requirements.txt
django==4.2
django-allauth==0.54.0
tweepy==4.10.0
facebook-sdk==3.1.0
pillow==10.0.0